backend/requirements.txt
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
requests==2.32.3
google-generativeai